progress toward more complex parser gen

+++ test/hello.zig
@@ -1,6 +1,4 @@
-
-
-fn main(argc: isize, argv: *mut u8) -> isize {
+fn main(argc: i32, argv: *mut u8) -> i32 {
     puts("Hello, world!\n");
     return 0;
 }
